==> notebooks/03_SynthControl.ipynb <==
## Synth Control Methods

**Abstract:**
One-sentence description

**Description:**
In the following cell, I...


%load_ext autoreload
%autoreload 2
%matplotlib inline
%config InlineBackend.figure_format = 'retina'

import pandas as pd
import numpy as np
from matplotlib import pyplot as plt

data = pd.read_csv("../data/processed/pca_transformed_0.85.csv", index_col=0)
data.head()

## 1987 年分水方案

import pandas as pd
from src.multi_synth import MultiSynth

features = ["PC1", "PC2", "PC3", "PC4", "PC5"]
Y_inputs = "Total water use"
province_include = [
    "Qinghai",
    "Gansu",
    "Ningxia",
    "Neimeng",
    "Shanxi",
    "Shaanxi",
    "Henan",
    "Shandong",
]
province_exclude = ["Jinji", "Beijing", "Sichuan"]


model_87 = MultiSynth(
    dataset=data,
    outcome_var=Y_inputs,
    time_var="Year",
    id_var="Province",
    treated_units=province_include,
    excluded_units=province_exclude,
    features=features,
    start=1979,
    end=1998,
    pen="auto",
    n_optim=100,
)

model_87.units
model_87.run_models(treated_time=1987, differenced=True, space_placebo=True)

## 1998年分水方案

model_98 = MultiSynth(
    dataset=data,
    outcome_var=Y_inputs,
    time_var="Year",
    id_var="Province",
    treated_units=province_include,
    excluded_units=province_exclude,
    features=features,
    start=1987,
    end=2008,
    pen="auto",
    n_optim=100,
)

model_98.units
model_98.run_models(treated_time=1998, differenced=True, space_placebo=True)

from src.plots import plot_pre_post

data = model_87.units["Henan"].result
plot_pre_post(data, 1987, Y_inputs)

model_87.export_to_pickle("../model/model_87.pkl")
model_98.export_to_pickle("../model/model_98.pkl")
